requirements.txt
# Trading Bot Dependencies
pyupbit>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
ta>=0.11.0
